+++ angular-architecture/src/app/commercial/commercial-routing.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 const routes: Routes = [
-  { path: 'products', loadChildren: 'app/commercial/products/products.module#ProductsModule' }
+  { path: 'products', loadChildren: './products/products.module#ProductsModule' }
 ];
 
 @NgModule({