+++ angular-architecture/src/app/main/internal-error/internal-error-page/internal-error-page.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-internal-error-page',
+  templateUrl: './internal-error-page.component.html',
+  styleUrls: ['./internal-error-page.component.scss']
+})
+export class InternalErrorPageComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}