+++ angular-architecture/src/app/main/main-routing.module.ts
@@ -2,10 +2,14 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
 const routes: Routes = [
-  { path: 'dashboard', loadChildren: 'app/main/dashboard/dashboard.module#DashboardModule' },
-  { path: 'page-not-found', loadChildren: 'app/main/not-found/not-found.module#NotFoundModule' },
+
+  { path: 'dashboard', loadChildren: './dashboard/dashboard.module#DashboardModule' },
+  { path: 'page-not-found', loadChildren: './not-found/not-found.module#NotFoundModule' },
+  { path: 'internal-error', loadChildren: './internal-error/internal-error.module#InternalErrorModule' },
+
   { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
   { path: '**', redirectTo: 'page-not-found', pathMatch: 'full' }
+
 ];
 
 @NgModule({